refactor: deduplicate AdvancedSettings and rename to SlippageSettings
