+++ apps/main/src/dex/components/PageRouterSwap/index.tsx
@@ -491,7 +491,6 @@ const QuickSwap = ({
         )}
         <DetailInfoSlippageTolerance
           maxSlippage={storeMaxSlippage}
-          testId="slippage-tolerance"
           stateKey={isStableswapRoute ? 'stable' : 'crypto'}
         />
       </div>
