+++ apps/main/src/lend/components/DetailInfoSlippageTolerance.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components'
-import AdvancedSettings from '@/lend/components/AdvancedSettings'
 import DetailInfo from '@ui/DetailInfo'
 import Icon from '@ui/Icon'
 import { formatNumber } from '@ui/utils'
+import { SlippageSettings } from '@ui-kit/features/slippage-settings'
 import { t } from '@ui-kit/lib/i18n'
 
 type Props = {
@@ -11,7 +11,7 @@ type Props = {
 
 const DetailInfoSlippageTolerance = ({ maxSlippage }: Props) => (
   <StyledDetailInfo label={t`Slippage tolerance:`}>
-    <StyledAdvancedSettings
+    <StyledSlippageSettings
       maxSlippage={maxSlippage}
       buttonIcon={
         <>
@@ -23,7 +23,7 @@ const DetailInfoSlippageTolerance = ({ maxSlippage }: Props) => (
   </StyledDetailInfo>
 )
 
-const StyledAdvancedSettings = styled(AdvancedSettings)`
+const StyledSlippageSettings = styled(SlippageSettings)`
   justify-content: flex-end;
 `
