+++ packages/curve-ui-kit/src/features/slippage-settings/index.ts
@@ -1,0 +1,1 @@
+export * from './SlippageSettings'
